+++ changelog.md
@@ -23,6 +23,7 @@
 - #390 Import: add a setting whether to send the completion email.
 - #276 Enable to set the time of day for start and end date of discounts.
 - #142 Web API do not take back in stock notifications into account.
+- #486 Add setting to capture payment when order status changes to *complete*.
 
 ### Bugfixes